requirements.txt
python-dotenv
pandas
numpy
pymongo
certifi
pymongo[srv]
scikit-learn
dill
pyyaml
xgboost
fastapi
uvicorn
ipykernel
apache-airflow
mlflow
dvc

-e .